push window_gs code to avoid cannot find the source code in the future

+++ application/src/main/java/common/bolts/transactional/gsw/GSWBolt.java
@@ -41,7 +41,7 @@ public abstract class GSWBolt extends TransactionalBolt {
                 return false;//not yet processed.
             }
             DataBox dataBox = ref.getRecord().getValues().get(1);
-            int read_result = Integer.parseInt(dataBox.getString().trim());
+            long read_result = Long.parseLong(dataBox.getString().trim());
             event.result.add(read_result);
         }
         return true;
